finish 2j comment

+++ src/main/resources/db/migration/V6__Add_Comment_table.sql
@@ -1,7 +1,7 @@
 create table comment(
-id bigint AUTO_INCREMENT primary key,
+id int AUTO_INCREMENT primary key,
 context varchar(1024) not null ,
-parent_id bigint not null ,
+parent_id int not null ,
 type int not null,
 commentator int not null,
 gmt_create bigint not null ,
